Get service worker/offline mode working! (see full commit message below)
For offline mode to work (and thus, for you to verify that it is working), you need 2 things:
-The web app is hosted on a specific domain, that is also specified in
the source code
-Built in production mode(for whatever reason) -- i.e. It won't work if
you use 'npm start'; you have to use 'npm build'

For the first one: The domain is stored in the solutions/.env file with
the key PUBLIC_URL. For now I left it as http://localhost:3000, so we
are expecting that the PWA will be hosted on port 3000. Once we get an
actual domain(theoretically) we would update PUBLIC_URL to e.g.
https://our-domain-for-google-solutions.com.

For the second note: This is where it gets slightly more complicated to
test the PWA, because you can't simply do "npm start" and it immediately
works. Instead, you must first run "npm build" which will do a
production build of the app and store it in the "solutions/build"
folder. This will take around 30 seconds or so (at least on my
computer). Then, you have to find a way to host these built files on the
port 3000 (or, once we have the domain, upload it to the web server).

The way I hosted it on 3000 was to download 'serve' from npm, i.e. 'npm
add --global serve'. This is a program that makes a webserver hosting
the files from a directory on your computer. Cd into the build folder
and run 'serve -l 3000' and navigate to localhost:3000 in your browser.
You should see the app! Now for the offline part-- this is how you know
it is working. Stop(Ctrl+C) the 'serve' so now its no longer hosting the
files. If yo refresh the page in the browser you will still see the app,
even though localhost:3000 no longer exists! Let me know if you have any
trouble testing this

For reference (Dont worry about the exact code too much because we
copied ours from create-react-app and its gonna be slightly different,
just useful for the concepts)
https://codingthesmartway.com/getting-started-with-progressive-web-apps-pwa/